# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/catalogs/{catalogId}": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the specified catalog.",
        "operationId": "Catalogs_Get",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceCatalog"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/child-catalog-infos": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets a list of child catalog info for the provided parent catalog identifier.",
        "operationId": "Catalogs_GetChildCatalogInfos",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The parent catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/timerange": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the specified catalog's time range.",
        "operationId": "Catalogs_GetTimeRange",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogTimeRange"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/availability": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the specified catalog availability.",
        "operationId": "Catalogs_GetAvailability",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "begin",
            "in": "query",
            "description": "Start date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 2
          },
          {
            "name": "end",
            "in": "query",
            "description": "End date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogAvailability"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/attachments": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets all attachments for the specified catalog.",
        "operationId": "Catalogs_GetAttachments",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/attachments/{attachmentId}/content": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the specified attachment.",
        "operationId": "Catalogs_GetAttachmentStream",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "description": "The attachment identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/{catalogId}/metadata": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the catalog metadata.",
        "operationId": "Catalogs_GetMetadata",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogMetadata"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Puts the catalog metadata.",
        "operationId": "Catalogs_PutMetadata",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "catalogMetadata",
          "description": "The catalog metadata to put.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogMetadata"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResourceCatalog": {
        "type": "object",
        "description": "A catalog is a top level element and holds a list of resources.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "The identifier."
          },
          "properties": {
            "type": "object",
            "description": "The map of properties.",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          },
          "resources": {
            "type": "array",
            "description": "The list of representations.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          }
        }
      },
      "Resource": {
        "type": "object",
        "description": "A resource is part of a resource catalog and holds a list of representations.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "The identifier."
          },
          "properties": {
            "type": "object",
            "description": "The map of properties.",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          },
          "representations": {
            "type": "array",
            "description": "The list of representations.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Representation"
            }
          }
        }
      },
      "Representation": {
        "type": "object",
        "description": "A representation is part of a resource.",
        "additionalProperties": false,
        "properties": {
          "dataType": {
            "description": "The data type.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/NexusDataType"
              }
            ]
          },
          "samplePeriod": {
            "type": "string",
            "description": "The sample period.",
            "format": "duration"
          },
          "kind": {
            "description": "The representation kind.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/RepresentationKind"
              }
            ]
          }
        }
      },
      "NexusDataType": {
        "type": "string",
        "description": "Specifies the Nexus data type.",
        "x-enumNames": [
          "UINT8",
          "UINT16",
          "UINT32",
          "UINT64",
          "INT8",
          "INT16",
          "INT32",
          "INT64",
          "FLOAT32",
          "FLOAT64"
        ],
        "enum": [
          "UINT8",
          "UINT16",
          "UINT32",
          "UINT64",
          "INT8",
          "INT16",
          "INT32",
          "INT64",
          "FLOAT32",
          "FLOAT64"
        ]
      },
      "RepresentationKind": {
        "type": "string",
        "description": "Specifies the representation kind.",
        "x-enumNames": [
          "Original",
          "Resampled",
          "Mean",
          "MeanPolarDeg",
          "Min",
          "Max",
          "Std",
          "Rms",
          "MinBitwise",
          "MaxBitwise",
          "Sum"
        ],
        "enum": [
          "Original",
          "Resampled",
          "Mean",
          "MeanPolarDeg",
          "Min",
          "Max",
          "Std",
          "Rms",
          "MinBitwise",
          "MaxBitwise",
          "Sum"
        ]
      },
      "CatalogInfo": {
        "type": "object",
        "description": "A structure for catalog information.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "The identifier."
          },
          "title": {
            "type": "string",
            "description": "The title."
          },
          "contact": {
            "type": "string",
            "description": "The contact.",
            "nullable": true
          },
          "license": {
            "type": "string",
            "description": "The license.",
            "nullable": true
          },
          "isReadable": {
            "type": "boolean",
            "description": "A boolean which indicates if the catalog is accessible."
          },
          "isWritable": {
            "type": "boolean",
            "description": "A boolean which indicates if the catalog is editable."
          },
          "isPublished": {
            "type": "boolean",
            "description": "A boolean which indicates if the catalog is published."
          },
          "isOwner": {
            "type": "boolean",
            "description": "A boolean which indicates if the catalog is owned by the current user."
          }
        }
      },
      "CatalogTimeRange": {
        "type": "object",
        "description": "A catalog time range.",
        "additionalProperties": false,
        "properties": {
          "begin": {
            "type": "string",
            "description": "The date/time of the first data in the catalog.",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "description": "The date/time of the last data in the catalog.",
            "format": "date-time"
          }
        }
      },
      "CatalogAvailability": {
        "type": "object",
        "description": "The catalog availability.",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "description": "The actual availability data.",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
      "CatalogMetadata": {
        "type": "object",
        "description": "A structure for catalog metadata.",
        "additionalProperties": false,
        "properties": {
          "contact": {
            "type": "string",
            "description": "The contact.",
            "nullable": true
          },
          "groupMemberships": {
            "type": "array",
            "description": "A list of groups the catalog is part of.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "overrides": {
            "description": "Overrides for the catalog.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ResourceCatalog"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Please enter 'Bearer {your JWT token}'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}